backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
playwright==1.40.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
lxml==4.9.3
supabase==2.0.2
langchain==0.0.350
langchain-openai==0.0.2